Nextjs migration (#95)
* deployment is almost there, a lot of configuration changes

* changes to package json
