+++ firebase.json
@@ -4,30 +4,49 @@
     "indexes": "firestore.indexes.json"
   },
   "hosting": {
-    "public": "build",
-    "ignore": [
-      "firebase.json",
-      "firestore.rules",
-      "storage.rules",
-      "firebase.indexes.json",
-      "**/.*",
-      "**/node_modules/**"
-    ],
-    "rewrites": [
-      {
-        "source": "/blog/prod-testing/*",
-        "function": "prerenderArticle"
-      },
-      {
-        "source": "**",
-        "destination": "/index.html"
-      }
-    ]
-  },
+      "public": "public/",
+      "ignore": [
+        "firebase.json",
+        "firestore.rules",
+        "storage.rules",
+        "firebase.indexes.json",
+        "**/.*",
+        "**/node_modules/**"
+      ],
+      "rewrites": [
+        {
+          "source": "/blog/prod-testing/*",
+          "function": "prerenderArticle"
+        },
+        {
+          "source": "**",
+          "function": "nextjs-server"
+        }
+      ],
+      "site": "startgnv-dev",
+      "cleanUrls": true
+    },
   "storage": {
     "rules": "storage.rules"
   },
   "functions": {
-    "predeploy": "npm --prefix \"$RESOURCE_DIR\" run build"
+    "predeploy": "npm run build-all",
+    "source": ".",
+    "ignore":[
+      "firebase.json",
+      "firestore.rules",
+      ".firebase/**",
+      ".firebaserc",
+      "firebase.indexes.json",
+      "storage.rules",
+      "**/.*",
+      "**/src/**",
+      "**/.vscode/**",
+      "**/node_modules/**",
+      "**/public/**",
+      "**/.next/cache/**",
+      "**/cypress/**",
+      "pages/**"
+    ]
   }
 }
