--- firepit-log.txt
@@ -1,24 +1,0 @@
-Welcome to firepit v1.1.0!
-Doing JSON parses for version checks at /snapshot/firepit/vendor/node_modules/firebase-tools/package.json
-JSONStream,abbrev,abort-controller,accepts,agent-base,agentkeepalive,aggregate-error,ajv,ansi-align,ansi-escapes,ansi-regex,ansi-styles,ansicolors,anymatch,aproba,archiver,archiver-utils,are-we-there-yet,argparse,array-flatten,arrify,as-array,asn1,assert-plus,ast-types,async,asynckit,aws-sign2,aws4,balanced-match,base64-js,basic-auth,basic-auth-connect,bcrypt-pbkdf,big-integer,bignumber.js,binary,binary-extensions,bl,blakejs,bluebird,body-parser,boxen,brace-expansion,braces,buffer,buffer-crc32,buffer-equal-constant-time,buffer-indexof-polyfill,buffers,bytes,cacache,cacheable-request,call-me-maybe,camelcase,cardinal,caseless,chainsaw,chalk,chardet,chokidar,chownr,ci-info,cjson,clean-stack,cli-boxes,cli-color,cli-cursor,cli-spinners,cli-table,cli-width,cliui,clone,clone-response,code-point-at,color,color-convert,color-name,color-string,colors,colorspace,combined-stream,commander,compare-semver,compress-commons,compressible,compression,concat-map,configstore,connect,console-control-strings,content-disposition,content-type,cookie,cookie-signature,core-util-is,cors,crc-32,crc32-stream,cross-env,cross-spawn,crypto-random-string,csv-streamify,d,dashdash,data-uri-to-buffer,debug,decompress-response,deep-extend,deep-freeze,deep-is,defaults,defer-to-connect,degenerator,delayed-stream,delegates,depd,destroy,dot-prop,dotenv,duplexer2,duplexer3,duplexify,ecc-jsbn,ecdsa-sig-formatter,ee-first,emoji-regex,enabled,encodeurl,encoding,end-of-stream,env-paths,err-code,es5-ext,es6-iterator,es6-symbol,es6-weak-map,escalade,escape-goat,escape-html,escape-string-regexp,escodegen,esprima,estraverse,esutils,etag,event-emitter,event-target-shim,events-listener,exegesis,exegesis-express,exit-code,exit-on-epipe,express,ext,extend,external-editor,extsprintf,fast-deep-equal,fast-json-stable-stringify,fast-levenshtein,fast-safe-stringify,fast-text-encoding,fast-url-parser,fecha,figures,file-uri-to-path,filesize,fill-range,finalhandler,firebase-tools,flat-arguments,fn.name,forever-agent,form-data,forwarded,fresh,fs-constants,fs-extra,fs-minipass,fs.realpath,fstream,ftp,gauge,gaxios,gcp-metadata,get-caller-file,get-stream,get-uri,getpass,glob,glob-parent,glob-slash,glob-slasher,global-dirs,google-auth-library,google-gax,google-p12-pem,got,graceful-fs,gtoken,har-schema,har-validator,has-ansi,has-flag,has-unicode,has-yarn,home-dir,http-cache-semantics,http-errors,http-proxy-agent,http-signature,https-proxy-agent,humanize-ms,iconv-lite,ieee754,import-lazy,imurmurhash,indent-string,infer-owner,inflight,inherits,ini,inquirer,install-artifact-from-github,ip,ip-regex,ipaddr.js,is-arrayish,is-binary-path,is-ci,is-extglob,is-fullwidth-code-point,is-glob,is-installed-globally,is-lambda,is-npm,is-number,is-obj,is-path-inside,is-promise,is-stream,is-stream-ended,is-typedarray,is-url,is-wsl,is-yarn-global,is2,isarray,isexe,isstream,jju,join-path,js-yaml,jsbn,json-bigint,json-buffer,json-parse-helpfulerror,json-ptr,json-schema,json-schema-traverse,json-stringify-safe,jsonfile,jsonparse,jsonwebtoken,jsprim,jwa,jws,keyv,kuler,latest-version,lazystream,leven,levn,listenercount,lodash,lodash._isnative,lodash._objecttypes,lodash._shimkeys,lodash.camelcase,lodash.defaults,lodash.difference,lodash.flatten,lodash.includes,lodash.isarguments,lodash.isboolean,lodash.isinteger,lodash.isnumber,lodash.isobject,lodash.isplainobject,lodash.isstring,lodash.keys,lodash.once,lodash.snakecase,lodash.toarray,lodash.union,lodash.values,log-symbols,logform,long,lowercase-keys,lru-cache,lru-queue,make-dir,make-fetch-happen,marked,marked-terminal,media-typer,memoizee,merge-descriptors,methods,mime,mime-db,mime-types,mimic-fn,mimic-response,minimatch,minimist,minipass,minipass-collect,minipass-fetch,minipass-flush,minipass-pipeline,minipass-sized,minizlib,mkdirp,morgan,ms,mute-stream,nan,nash,negotiator,netmask,next-tick,nice-try,node-emoji,node-fetch,node-forge,node-gyp,nopt,normalize-path,normalize-url,npmlog,number-is-nan,oauth-sign,object-assign,object-hash,on-finished,on-headers,once,one-time,onetime,open,openapi3-ts,optionator,ora,os-tmpdir,p-cancelable,p-defer,p-map,pac-proxy-agent,pac-resolver,package-json,parseurl,path-is-absolute,path-key,path-to-regexp,performance-now,picomatch,portfinder,prelude-ls,prepend-http,printj,process-nextick-args,progress,promise-breaker,promise-inflight,promise-retry,protobufjs,proxy-addr,proxy-agent,proxy-from-env,psl,pump,punycode,pupa,qs,range-parser,raw-body,rc,re2,readable-stream,readdir-glob,readdirp,redeyed,registry-auth-token,registry-url,request,require-directory,responselike,restore-cursor,retry,retry-request,rimraf,router,rsvp,run-async,rxjs,safe-buffer,safer-buffer,semver,semver-diff,send,serve-static,set-blocking,setimmediate,setprototypeof,shebang-command,shebang-regex,signal-exit,simple-swizzle,smart-buffer,socks,socks-proxy-agent,source-map,sprintf-js,sshpk,ssri,stack-trace,statuses,stream-shift,string-length,string-width,string_decoder,strip-ansi,strip-json-comments,superstatic,supports-color,supports-hyperlinks,tar,tar-stream,tcp-port-used,term-size,text-hex,through,through2,timers-ext,tmp,to-readable-stream,to-regex-range,toidentifier,tough-cookie,toxic,traverse,triple-beam,tslib,tunnel-agent,tweetnacl,tweetsodium,type,type-check,type-fest,type-is,typedarray-to-buffer,unique-filename,unique-slug,unique-string,universal-analytics,universalify,unpipe,unzipper,update-notifier,uri-js,url-join,url-parse-lax,util-deprecate,utils-merge,uuid,valid-url,vary,verror,wcwidth,which,wide-align,widest-line,winston,winston-transport,word-wrap,wrap-ansi,wrappy,write-file-atomic,ws,xdg-basedir,xregexp,xtend,y18n,yallist,yaml,yargs,yargs-parser,zip-stream,@apidevtools,@dabh,@google-cloud,@grpc,@jsdevtools,@npmcli,@opentelemetry,@protobufjs,@sindresorhus,@szmarczak,@tootallnate,@types
-Installed ft@9.16.0 and packaged ft@9.16.0
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/lib/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /snapshot/firepit/node_modules/npm/bin/npm-cli
-Found npm/bin/npm-cli install.
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/lib/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /snapshot/firepit/node_modules/npm/bin/npm-cli
-Found npm/bin/npm-cli install.
-Runtime binaries created.
-/usr/local/bin/firebase
-/usr/local/bin/firebase,/snapshot/firepit/firepit.js,apps:sdkconfig,web,--json
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/lib/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /snapshot/firepit/node_modules/npm/bin/npm-cli
-Found npm/bin/npm-cli install.
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/lib/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /Users/admin/.cache/firebase/tools/node_modules/npm/bin/npm-cli
-Checking for npm/bin/npm-cli install at /snapshot/firepit/node_modules/npm/bin/npm-cli
-Found npm/bin/npm-cli install.
-ShellJSInternalError: ENOENT: no such file or directory, stat '/Users/admin/.cache/firebase/runtime/node'
