+++ functions/.gitignore
@@ -9,3 +9,4 @@ lib/
 typings/
 
 node_modules/
+next
