+++ next.config.js
@@ -5,5 +5,15 @@ module.exports = {
           config.resolve.fallback.module = false;
       }
       return config;
+    },
+    typescript: {
+      // !! WARN !!
+      // Dangerously allow production builds to successfully complete even if
+      // your project has type errors.
+      // !! WARN !!
+      ignoreBuildErrors: true,
+    },
+    env: {
+      FIREBASE_PROJECT_ID: "startgnv-dev"
     }
   };
