+++ package.json
@@ -3,23 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "run-p -r start:next start:firebase",
-    "start:firebase": "firebase serve",
-    "start:e2e": "env-cmd -f ./.env.test yarn start",
-    "start:dev": "run-p getConfig start:firebase dev",
+    "start": "run-p -r serve next",
+    "start-dev": "run-p serve dev",
+    "next": "next start",
     "dev": "next dev",
-    "build": "npm run getConfig && next build",
-    "start:next": "npm run getConfig && next start",
+    "serve": "npm run getConfig && firebase serve",
     "lint": "next lint",
-    "getConfig": "firebase apps:sdkconfig web --json > ./firebase-config.json"
+    "getConfig": "firebase apps:sdkconfig web --json > ./firebase-config.json",
+    "build-all": "npm run build-next && npm run build-firebase",
+    "build-firebase": " cd \"functions\" && npm install && npm run build",
+    "build-next": "npm install && next build",
+    "predeploy": "npm run clean && npm run build-all"
   },
   "devDependencies": {
     "babel-plugin-styled-components": "latest",
     "cypress": "^4.12.1",
     "cypress-firebase": "^1.4.3",
     "env-cmd": "^10.1.0",
-    "firebase": "^8.2.6",
-    "firebase-admin": "^9.0.0",
     "npm-run-all": "^4.1.5",
     "typescript": "^3.9.7"
   },
@@ -40,10 +40,13 @@
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.8.4",
     "email-validator": "^2.0.4",
-    "firebase-functions": "^3.13.1",
+    "firebase": "^8.10.0",
+    "firebase-admin": "^9.11.1",
+    "firebase-functions": "^3.15.4",
     "html-to-draftjs": "^1.4.0",
     "html-to-react": "^1.4.2",
     "husky": "^1.3.1",
+    "immutable": "^4.0.0-rc.14",
     "lint-staged": "^8.1.5",
     "moment": "^2.24.0",
     "next": "11.1.0",
@@ -85,5 +88,9 @@
       "pre-commit": "lint-staged"
     }
   },
-  "proxy": "http://localhost:5000"
+  "proxy": "http://localhost:5000",
+  "engines": {
+    "node": "14"
+  },
+  "main": "server.js"
 }
