+++ src/firebase.js
@@ -1,5 +1,6 @@
 import firebase from 'firebase';
 import firebaseConfig from 'firebase-config.json';
+import 'firebase/storage';
 
 const CLIENT_CONFIG = firebaseConfig.result.sdkConfig;
